requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.6.15.1
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.50.0
django-ckeditor==6.5.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
drf-yasg==1.21.4
filelock==3.8.0
flake8==5.0.4
flake8-black==0.3.3
gunicorn==20.1.0
identify==2.5.8
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.1
packaging==21.3
pathspec==0.10.1
phonenumbers==8.12.54
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.7
whitenoise==5.3.0
django-rest-passwordreset==1.3